# Add streaming match generation