This PR proposes "streaming match generation" for `GrandCypher`. It depends on the changes in this corresponding PR https://github.com/aplbrain/grandiso-networkx/pull/39.

This change is designed to improve the performance of large openCypher queries. It employs a streaming match generation strategy to iteratively process each result. 

For example, with this code snippet:

```python
import time

import networkx as nx

from grandcypher import GrandCypher

host = nx.DiGraph()

start = time.time()
size = 1000000
for x in range(size):
    host.add_node(x, name=x)
    host.add_edge(x, x + 1)

print(f"Built graph in {time.time() - start}s")

qry = """
MATCH (A) -[]-> (B) -[]-> (C) -[]-> (D)
RETURN A.name, B.name, C.name, D.name
LIMIT 1
"""

start = time.time()
res = GrandCypher(host).run(qry)
print(f"Query took {time.time() - start}s")
print(res)
```

The current code base takes:

```
Built graph in 3.705066442489624s
Query took 86.24252390861511s
{'A.name': [0], 'B.name': [1], 'C.name': [2], 'D.name': [3]}
```

With this PR, it takes:

```
Built graph in 2.572826623916626s
Query took 0.003414630889892578s
{'A.name': [0], 'B.name': [1], 'C.name': [2], 'D.name': [3]}
```

While this query on the surface seems ridiculous, it's not too different than:

```sql
SELECT * from table LIMIT 1
```

For situations where you may just want to sample a few rows. 

Here are a couple additional examples:

```python
qry = """
MATCH (A{name: 0}) -[]-> (B) -[]-> (C) -[]-> (D)
RETURN A.name, B.name, C.name, D.name
LIMIT 1
"""
```

Current:
```
Built graph in 2.583895444869995s
Query took 0.6473803520202637s
{'A.name': [0], 'B.name': [1], 'C.name': [2], 'D.name': [3]}
```

With this PR:
```
Built graph in 3.0622048377990723s
Query took 0.003484010696411133s
{'A.name': [0], 'B.name': [1], 'C.name': [2], 'D.name': [3]}
```

```python
qry = """
MATCH (A) -[]-> (B{name: 1}) -[]-> (C) -[]-> (D)
RETURN A.name, B.name, C.name, D.name
LIMIT 1
"""
```

Current:
```
Built graph in 2.8512825965881348s
Query took 25.05796718597412s
{'A.name': [0], 'B.name': [1], 'C.name': [2], 'D.name': [3]}
```

With this PR:
```
Built graph in 2.822657346725464s
Query took 0.0033731460571289062s
{'A.name': [0], 'B.name': [1], 'C.name': [2], 'D.name': [3]}
```

```python
qry = """
MATCH (A) -[]-> (B) -[]-> (C) -[]-> (D)
WHERE D.name = 100
RETURN A.name, B.name, C.name, D.name
LIMIT 1
"""
```

Current:
```
Built graph in 3.100694179534912s
Query took 85.86176109313965s
{'A.name': [97], 'B.name': [98], 'C.name': [99], 'D.name': [100]}
```

With this PR:
```
Built graph in 2.6066715717315674s
Query took 0.013916015625s
{'A.name': [97], 'B.name': [98], 'C.name': [99], 'D.name': [100]}
```

This change also adds an optional default limit that can be applied to all queries. It's worth noting that this only improves performance when a query has results. For queries without results, it still goes through the whole graph looking for results. I thought about adding a maximum iterations parameter but didn't include that with this PR.