@davidmezzetti 2024-02-13
That's correct, this change works with the current version of `grandiso`. But as you said, it wouldn't benefit from the short circuiting logic without the changes from `grandiso`.