@j6k4m8 2024-02-13
Ok I've thought about it a lot and I think the "right" answer is to NOT bump the version dep on grandiso in this library, since it's a nonbreaking and backwards compatible change and I'd rather avoid the possibility of adding incompatibilities into existing projects; but I will merge this presently and anyone who installs from now on should get the latest of both, and enjoy the huge performance bump from your work!! 


Thank you again! Excited to play around with this some more :)