@davidmezzetti 2024-02-13
Sounds like a good approach to me!